Update `BloodSugarMeasurements` DB schema
- Change `BloodSugarReading#value` type from `Int` to `Float`

+++ app/src/main/java/org/simple/clinic/bloodsugar/BloodSugarReading.kt
@@ -4,7 +4,7 @@ import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-data class BloodSugarReading(val value: Int, val type: BloodSugarMeasurementType) : Parcelable {
+data class BloodSugarReading(val value: Float, val type: BloodSugarMeasurementType) : Parcelable {
   val isHigh: Boolean
     get() {
       return when (type) {
